mcp-planner/requirements.txt
mcp[cli]
fastapi
uvicorn[standard]
python-jose[cryptography]
sqlmodel
